rearrange imports so numbered files aren't imported
